03-sheet-2-mysql/requirements.txt
gspread
mysql-connector-python
python-dotenv
